feat: add polish macro for extra substitutions
Coming from Hugo, the markdown parser substitutes -- for endash and ---
for emdash. This replicates that functionality.

In addition, without actually implementing a custom Tera filter, we
achieve a hacky basic set of emoji support by just running a set of
replacements via the macro. Only a small subset is included now, but it
can easily be added to. Ideally, this will be added to Zola or Tera in
the future.