refactor: preconnect google fonts
We can mitigate some of this performance hit of downloading the CSS and
then the fonts from two different domains. The first domain (for the
CSS) should be fairly high up your index.html so hopefully will be seen
early enough, but the second domain is not seen until later. However we
know what that domain will be (fonts.gstatic.com), so we can use a
preconnect resource hint to ask the browser to open the connection in
advance to save some of that second connection time later:

Reference:
    https://www.tunetheweb.com/blog/should-you-self-host-google-fonts/