Fix bug in title tag when subtitle missing
If `subtitle` is not set in extra then an error will occur when
rendering the index template (specifically in the <title> tag). This
change checks for the presence of a subtitle before attempting to render
it.

Additionally, if the subtitle was the empty string the title would be
rendered with an additional | character after the name (e.g. `mysite
|`). This change removes that character when the subtitle is not
present.