fix(index.html): strip tags from page summary before trunc (#6)
Strip tags from the page summary before truncating the content to be
disapled on the front page. If the content truncates at the middle of an
open tag, i.e. does not include a closing tag in the 280 chars, the Read
more anchor tag is corrupted.

Documented in https://github.com/getzola/zola/issues/1149

Co-authored-by: Prateek Nischal <PNischal@walmartlabs.com>